refactor: extract order logic into service object
This should make it easier to test, hopefully
